combine ICD codes

These function implement combination of lists or vectors of codes, while preserving ICD classes. Base R c just drops all user defined classes and casts down to lowest common denominator, e.g. if mixing numbers and characters. No attempt here to catch all possible combinations of feeding in mixed ICD types and other types. Let R do what it normally does, but just try to keep classes of the first item in the list.

# S3 method for icd9
c(..., warn = FALSE)

# S3 method for icd10
c(..., warn = FALSE)

Arguments

...
elements to combine

warn
single logical value, if TRUE, will give warnings when incompatible types are combined using c

See also

Examples

# Care with the following:
c(as.icd9("E998"), as.icd10("A10"))

#> [1] "E998" "A10" 

# which results in both codes sharing the 'icd9' class.
# ICD-10 codes
(a <- as.icd10("A100SSX"))

#> [1] "A100SSX"

(b <- as.icd10("Z999A"))

#> [1] "Z999A"

c(a, b)

#> [1] "A100SSX" "Z999A"  

c(as.icd_short_diag(a), as.icd_short_diag(b))

#> [1] "A100SSX" "Z999A"  

(d <- as.icd10("A10.0SSX"))

#> [1] "A10.0SSX"

(e <- as.icd10("Z99.9A"))

#> [1] "Z99.9A"

c(d, e)

#> [1] "A10.0SSX" "Z99.9A"  

c(as.icd_decimal_diag(d), as.icd_decimal_diag(e))

#> [1] "A10.0SSX" "Z99.9A"  

# warn when mixing attribute types
suppressWarnings(
  c(as.icd_short_diag(a), as.icd_decimal_diag(e))
)

#> [1] "A100SSX" "Z99.9A"