Calculate Charlson scores from precomputed Charlson comorbidities

Calculate Charlson scores from precomputed Charlson comorbidities, instead of directly from the ICD codes. This is useful if the comorbidity calculation is time consuming. Commonly, both the Charlson comorbidities and the Charlson scores will be calculated, and this function provides just that second step.

charlson_from_comorbid(
  x,
  visit_name = NULL,
  hierarchy = FALSE,
  scoring_system = c("original", "charlson", "quan")
)

Arguments

x
data.frame or matrix, typically the output of a comorbidity calculation which uses the Charlson categories, e.g. comorbid_quan_deyo

visit_name
The name of the column in the data frame which contains the patient or visit identifier. Typically this is the visit identifier, since patients come leave and enter hospital with different ICD-9 codes. It is a character vector of length one. If left empty, or NULL, then an attempt is made to guess which field has the ID for the patient encounter (not a patient ID, although this can of course be specified directly). The guesses proceed until a single match is made. Data frames may be wide with many matching fields, so to avoid false positives, anything but a single match is rejected. If there are no successful guesses, and visit_id was not specified, then the first column of the data frame is used.

hierarchy
single logical value, default is FALSE. If TRUE, will drop DM if DMcx is present, etc.

scoring_system
One of original, charlson, or quan. The first two will give the original Charlson weights for each comorbidity, whereas quan uses the updated weights from Quan 2011.