convert comorbidity matrix to data frame

convert matrix of comorbidities into data frame, preserving visit_name information

comorbid_df_to_mat(x, visit_name = get_visit_name(x),
  stringsAsFactors = getOption("stringsAsFactors"))

icd_comorbid_df_to_mat(...)

icd_comorbid_mat_to_df(...)

Arguments

x
data frame, with a visit_name column (not necessarily first), and other columns with flags for comorbidities, as such column names are required.

visit_name
The name of the column in the data frame which contains the patient or visit identifier. Typically this is the visit identifier, since patients come leave and enter hospital with different ICD-9 codes. It is a character vector of length one. If left empty, or NULL, then an attempt is made to guess which field has the ID for the patient encounter (not a patient ID, although this can of course be specified directly). The guesses proceed until a single match is made. Data frames may be wide with many matching fields, so to avoid false positives, anything but a single match is rejected. If there are no successful guesses, and visit_id was not specified, then the first column of the data frame is used.

stringsAsFactors
Single logical value, describing whether the resulting data frame should have strings, e.g. visit_id converted to factor. Default is to follow the current session option. This is identical to the argument used in, among other base functions as.data.frame.

...
arguments passed on to other functions

Deprecated function names

Future versions of icd will drop the icd_ prefix. For example, charlson should be used in favor of icd_charlson. To distinguish icd function calls, consider using the prefix icd:: instead, e.g., icd::charlson. Functions which specifically operate on either ICD-9 or ICD-10 codes or their sub-types will retain the prefix. E.g. icd9_comorbid_ahrq. icd specific classes also retain the prefix, e.g., icd_wide_data.

Examples

longdf <- icd_long_data(
            visit = c("a", "b", "b", "c"),
            icd9 = c("441", "4240", "443", "441")
            )
cmbdf <- icd9_comorbid_elix(longdf, return_df = TRUE)
class(cmbdf)

#> [1] "data.frame"

rownames(cmbdf)

#> [1] "1" "2" "3"

mat.out <- comorbid_df_to_mat(cmbdf)
stopifnot(is.matrix(mat.out))
mat.out[, 1:4]

#>     CHF Arrhythmia Valvular  PHTN
#> a FALSE      FALSE    FALSE FALSE
#> b FALSE      FALSE     TRUE FALSE
#> c FALSE      FALSE    FALSE FALSE