Convert Decimal format ICD codes to short format

This usually just entails removing the decimal point, but also does some limited validation and tidying up.

decimal_to_short(x)

# S3 method for icd9
decimal_to_short(x)

# S3 method for icd10
decimal_to_short(x)

# S3 method for icd10cm
decimal_to_short(x)

# S3 method for default
decimal_to_short(x)

Arguments

x
ICD codes

Methods (by class)

icd9: convert ICD-9 codes from decimal to short format

icd10: convert ICD-10 codes from decimal to short format

icd10cm: convert ICD-10-CM codes from decimal to short format

default: Guess ICD version and convert decimal to short format

See also