Explain ICD-9 and ICD-10 codes in English from decimal (123.45 style), Tabulates the decimal format alongside converted non-decimal format.

Output is ordered in the same order as the input. A logical column ismajor indicates if the code is a parent Category.

explain_table(...)

# S3 method for default
explain_table(x, short_code = guess_short(x),
  condense = FALSE, brief = TRUE, warn = TRUE, ...)

# S3 method for icd9
explain_table(...)

# S3 method for icd10
explain_table(...)

# S3 method for icd9cm
explain_table(x, short_code = guess_short(x),
  condense = FALSE, brief = TRUE, warn = TRUE, ...)

# S3 method for icd10cm
explain_table(x, short_code = guess_short(x),
  condense = FALSE, brief = TRUE, warn = TRUE, ...)

# S3 method for icd10who
explain_table(x, short_code = guess_short(x),
  condense = FALSE, brief = TRUE, warn = TRUE, ...)

Arguments

...
arguments passed on to other functions

x
vector or other structure of ICD codes to explain in human language

short_code
single logical value which determines whether the ICD-9 code provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this is guessed from the input data.

condense
single logical value which indicates whether to condense the given set of ICD-9 codes by replacing subsets of codes with 'parent' codes which exactly encompass certain subsets. E.g. If all cholera diagnoses are provided, only '001 - Cholera' needs to be displayed, not all sub-types.

brief
single logical value, default is FALSE. If TRUE, the short description from the canonical CMS descriptions will be used, otherwise the long description is used.

warn
single logical value, default is TRUE, meaning that codes which do not correspond to diagnoses, or to three-digit codes, will trigger a warning.

Value

data frame with fields for ICD-9 code, name and description. The ordering is in the same order as input, including rows filled with NA for invalid input codes

Details

If the code is both a valid ICD9 and a ICD10 the output will default the descriptions to ICD10. The code would otherwise have to be explicitly cast to get ICD9 descriptions.

A column for source year may be added in the future. Other changes may occur this new feature gets testing and use.

If the input x is of mixed type it will choose to convert by

Methods (by class)

default: explaining ICD codes from a character vector, guessing ICD version

icd9: from vector of ICD-9 codes.

icd10: from vector of ICD-10 codes.

icd9cm: explain character vector of ICD1-10-CM codes

icd10cm: explain character vector of ICD1-10-CM codes

icd10who: explain character vector of ICD1-10-CM codes