Filter ICD codes by validity.

Filters a data.frame of patients for valid or invalid ICD codes

filter_valid(
  x,
  icd_name = get_icd_name(x),
  short_code = guess_short(.subset2(x, icd_name)),
  invert = FALSE,
  ...
)

filter_invalid(
  x,
  icd_name = get_icd_name(x),
  short_code = guess_short(x[[icd_name]]),
  invert = FALSE
)

icd9_filter_valid(
  x,
  icd_name = get_icd_name(x),
  short_code = guess_short(x[[icd_name]]),
  invert = FALSE
)

icd10_filter_valid(
  x,
  icd_name = get_icd_name(x),
  short_code = guess_short(x[[icd_name]]),
  invert = FALSE
)

icd9_filter_invalid(
  x,
  icd_name = get_icd_name(x),
  short_code = guess_short(x[[icd_name]]),
  invert = FALSE
)

icd10_filter_invalid(
  x,
  icd_name = get_icd_name(x),
  short_code = guess_short(x[[icd_name]]),
  invert = FALSE
)

Arguments

x
a data.frame containing a column of ICD codes

icd_name
The name of the column in the data.frame which contains the ICD codes. This is a character vector of length one. If it is NULL, icd9 will attempt to guess the column name, looking for progressively less likely possibilities until it matches a single column. Failing this, it will take the first column in the data frame. Specifying the column using this argument avoids the guesswork.

short_code
single logical value which determines whether the ICD-9 code provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this is guessed from the input data.

invert
Single logical value. Returns the inverse of the result. E.g. if seeking valid ICD-9 codes, the invalid ones are returned.

...
arguments passed to the class-specific functions

Functions

filter_invalid: Filter invalid rows from data frame of patients with ICD codes. This can also be achieved with filter_valid and invert = TRUE

icd9_filter_valid: Filter data frame for valid ICD codes