Select only defined ICD codes

Return only those codes which are heading or leaf (billable), specifying whether codes are all short-form or all decimal-form

get_defined(x, short_code = guess_short(x), billable = FALSE)

icd_get_defined.icd9(...)

icd_get_defined(...)

Arguments

x
input vector or factor, possibly with an ICD class

short_code
logical value, whether short-form ICD code

billable
single logical value, whether to limit return codes also by whether they are billable, i.e. leaf nodes. This is really only designed for use with ICD-9-CM, ICD-10-CM etc, since the WHO versions are not designed for billing, but for public health and death reporting.

...
arguments passed on to other functions

Deprecated function names

Future versions of icd will drop the icd_ prefix. For example, charlson should be used in favor of icd_charlson. To distinguish icd function calls, consider using the namespace prefix icd:: instead, e.g., icd::charlson. Functions which specifically operate on either ICD-9 or ICD-10 codes or their subtypes will retain the prefix. E.g. icd9_comorbid_ahrq. icd specific classes also retain the prefix, e.g., icd_wide_data.