Guess whether codes are short_code or decimal_code

The goal is to guess whether codes are short_code or decimal_code form. Currently condense works, but not with the icd look-up table currently in use. Of note, validation is a bit different here, since we don't know the type until after we guess. We could look for where both short_code and long are invalid, and otherwise assume valid, even if the bulk are short_code. However, it may be more useful to check validity after the guess.

guess_short(x_, short_code = NULL, n = 1000L, icd_name = NULL)

Value

single logical value, TRUE if input data are predominantly short_code type. If there is some uncertainty, then return NA.

Details

Very quick heuristic, ploughs through ten million codes in less than one second and will stop more quickly if it finds a '.'.