AHRQ ICD-10-PCS categories

The AHRQ has categorized each of the ICD-10-PCS (Procedure Codes) into one of four groups: minor diagnostic, minor therapeutic, major diagnostic or major therapeutic. This mapping can be used to get the type(s) of procedure(s) performed on a patient from a data.frame of patients and associated procedure codes in 'long' format. See the ICD-10 vignette for an example.

Details

Currently there is no specific comorbidity function to use this data, so the generic comorbid icd9_comorbid icd10_comorbid should be used, and this data specified as the map.

See also

Examples

icd10_map_ahrq_pcs[["Major Diagnostic"]][1:5]

#> [1] "00900ZX" "00910ZX" "00920ZX" "00930ZX" "00940ZX"

icd10_map_ahrq_pcs[["Minor Therapeutic"]][1:5]

#> [1] "0020X0Z" "0020XYZ" "002EX0Z" "002EXYZ" "002UX0Z"