United States and Belgium ICD-10-CM

The public domain modified ICD-10 classification as published in the public domain by the US CDC. Currently this has a slightly different structure to icd9cm_hierarchy because the published data helpfully has a leaf flag indicating whether a code is a billable leaf node, or a code higher in the hierarchy which nevertheless will have a description.

Source

http://www.cdc.gov/nchs/icd/icd10cm.htm

Details

Format: data frame, with columns for code, leaf status (0 or 1), short and long descriptions.

Editions

There are annual revisions to this data in the US.

References

https://www.cms.gov/Medicare/Coding/ICD10/ https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2018-ICD-10-PCS-Tables-And-Index.zip https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2018-ICD-10-PCS-Order-File.zip https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2017-PCS-Code-Tables.zip https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2017-PCS-Long-Abbrev-Titles.zip https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2016-Code-Descriptions-in-Tabular-Order.zip https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2015-code-descriptions.zip https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2015-tables-index.zip https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2015-Code_Tables-and-Index.zip https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2015-PCS-long-and-abbreviated-titles.zip https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2014-ICD10-Code-Descriptions.zip https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2014-ICD10-Code-Tables-and-Index.zip https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2014-Code-Tables-and-Index.zip https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2014-PCS-long-and-abbreviated-titles.zip

See also