Convert mjr and mnr vectors to single code

Convert mjr and mnr vectors to single code

icd9MajMinToCode(mjr, mnr, isShort)

Arguments

isShort
Deprecated. Single logical value which determines whether the ICD-9 code provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this is guessed from the input data. Use short_code instead.

major
character vector of 'major' part of ICD-9 codes, i.e. that part which falls before the decimal point, in decimal notation. (In five digit notation, the 'major' part is be the first three characters (with leading zeroes), and includes V or E prefix.

minor
character vector of 'minor' part of ICD-9 codes, i.e. that part which falls after the decimal point, in decimal notation. (In 'short' five digit notation, the 'major' part is the first three characters including leading zeroes which may be inferred; the last two characters represent the 'minor' part.)

Value

Character vector