Quan adaptation of Deyo/Charlson comorbidities

Derived automatically from the SAS code used in the original publication. According to the referenced study, this provides the best predictor of in-patient to <30d mortality. Of note, Deyo drops the distinction between leukemia, lymphoma and non-metastatic cancer. As far as I have looked into this, in the rare cases where someone had two or three of leukemia, lymphoma and non-metastatic cancer, the Quan adaptation would give a lower Charlson score than the original scheme. The original Deyo Charlson to ICD-9-CM groups does include distinct categories for these things.

Format

list of character vectors, each named by co-morbidity

References

Quan, Hude, Vijaya Sundararajan, Patricia Halfon, Andrew Fong, Bernard Burnand, Jean-Christophe Luthi, L. Duncan Saunders, Cynthia A. Beck, Thomas E. Feasby, and William A. Ghali. "Coding Algorithms for Defining Comorbidities in ICD-9-CM and ICD-10 Administrative Data." Medical Care 43, no. 11 (November 1, 2005): 1130-39. http://www.ncbi.nlm.nih.gov/pubmed/16224307 http://web.archive.org/web/20110225042437/http://www.chaps.ucalgary.ca/sas

See also