Generate ICD to HCC Crosswalks from CMS

The ICD/HCC mappings were obtained from CMS Medicare Risk Adjustors. Due to the complex file structure of the original data (many nested zip files), they have been organized in the folder data/icd_hcc_rawdata/ available in the source repository but not the installed package. This function creates a data file containing ICD-9/10 to CC crosswalks.

icd9_parse_cc(save_data = FALSE)

Arguments

save_data
logical whether to save the result in the source tree. Defaults to FALSE.

condense
Deprecated. The map is not condensed by default. If required, it can reduced to a smaller set of codes using icd_condense.

Value

invisibly returns the list of vectors, where a co-morbidity name is associated with a character vector of ICD-9 codes.