Read the ICD-9-CM description data as provided by the Center for Medicaid Services (CMS).

The full ICD-9 specification is in an RTF file, but CMS also distributes a space-separated text file with just the definitions for 'defined', 'billable' codes. Note that this canonical data doesn't specify non-diagnostic higher-level codes, just the specific diagnostic 'child' codes.

icd9_parse_leaf_desc_ver(version = icd9cm_latest_edition(),
  save_data = TRUE, offline = TRUE)

Arguments

version
character vector of length one containing the ICD-9 version, e.g. "32" which is the default.

save_data
single logical value, if TRUE will save the generated data

offline
single logical, if TRUE then don't pull the file from internet, only return path and file name if the file already exists in the raw data directory. This is helpful for testing without using the internet.

path
Absolute path in which to save parsed data

Value

invisibly return the result

Details

The file can be pulled from the zip files on the CMS web site or from within the package.