ICD-9-CM diagnosis codes including leaf nodes and branch names up to the three-digit codes.

Unlike get_icd9cm2014_leaf and friends, these data frames contain the full structure of the ICD-9-CM scheme up to the three-digit codes. Unlike the equivalent ICD-10-CM data frames (e.g., icd10cm2019), they do not have columns indicating billable/leaf status, or chapter designations.

Source

http://wonder.cdc.gov/wonder/sci_data/codes/icd9/type_txt/icd9cm.asp

Rich text descriptions here: http://www.cdc.gov/nchs/icd/icd9cm.htm http://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html This page has versions 23 to 32 (2005 to 2014). At present, only the 2014 data is included in this package.

http://wonder.cdc.gov/wonder/sci_data/codes/icd9/type_txt/icd9abb.asp

http://wonder.cdc.gov/wonder/sci_data/codes/icd9/type_txt/icd9cm.asp

http://wonder.cdc.gov/wonder/sci_data/codes/icd9/type_txt/icdcm.asp

http://wonder.cdc.gov/wonder/sci_data/codes/icd9/type_txt/icd9abb.asp

Details

icd9cm_hierarchy is the deprecated name, which currently points to the final ICD-9-CM 2014 release. 2011 -- 2014 are in fact identical. These are generated dynamically by parsing an awkward RTF file, and only the 2011 -- 2014 data has been tested and used thoroughly.

Format: data frame