OAuth2 authentication to ICD API

OAuth2 authentication to ICD API

Usage

icd_endpoint(
  authorize = NULL,
  access = "https://icdaccessmanagement.who.int/connect/token"
)

icd_oauth_app(
  appname = "icd_app",
  key = "6fc8a1e4-4da9-43a8-bd0c-c164c0cb0ebd_3c7e272e-2f4b-46de-b127-df7454e36be8",
  secret = "OGIly8mTatQ2ILuhBTbviIp2FofWiQR3fj4HaPiGseE="
)

icd_token(
  endpoint = icd_endpoint(),
  app = icd_oauth_app(),
  scope = "icdapi_access",
  as_header = TRUE,
  client_credentials = TRUE,
  ...
)

Arguments

authorize

url to send client to for authorisation. Set to NULL if not needed.

access

url used to exchange unauthenticated for authenticated token.

appname

Name of the application. This is not used for OAuth, but is used to make it easier to identify different applications.

key

Consumer key, also sometimes called the client ID.

secret

Consumer secret, also sometimes called the client secret. Despite its name, this does not necessarily need to be protected like a password, i.e. the user still has to authenticate themselves and grant the app permission to access resources on their behalf.

endpoint

OAuth endpoint for the ICD API, created by `icd_endpoint()`

app

An OAuth consumer application, created by `icd_oauth_app()` by default.

scope

ICD API scope to request. Default is "icdapi_access".

as_header

If TRUE, the default, configures the token to add itself to the bearer header of subsequent requests. If FALSE, configures the token to add itself as a url parameter of subsequent requests.

client_credentials

Default to FALSE. Set to TRUE to use Client Credentials Grant instead of Authorization Code Grant. See https://www.rfc-editor.org/rfc/rfc6749#section-4.4.

...

Other parameters/arguments to be passed onto `httr::oauth2.0_token()`

Value

For `icd_token()`, a Token2.0 reference class (RC) object.

Examples

icd_endpoint()
#> <oauth_endpoint>
#>  access: https://icdaccessmanagement.who.int/connect/token
icd_token()
#> <Token>
#> <oauth_endpoint>
#>  access: https://icdaccessmanagement.who.int/connect/token
#> <oauth_app> icd_app
#>   key:    6fc8a1e4-4da9-43a8-bd0c-c164c0cb0ebd_3c7e272e-2f4b-46de-b127-df7454e36be8
#>   secret: <hidden>
#> <credentials> access_token, expires_in, token_type
#> ---
icd_oauth_app()
#> <oauth_app> icd_app
#>   key:    6fc8a1e4-4da9-43a8-bd0c-c164c0cb0ebd_3c7e272e-2f4b-46de-b127-df7454e36be8
#>   secret: <hidden>