Check whether ICD-9 codes exist

This is different from syntactic validity: it looks up codes in the canonical lists published by the CMS.

is_defined(x, short_code = guess_short(x), ...)

# S3 method for icd9
is_defined(x, short_code = guess_short(x),
  billable = FALSE, leaf = billable, ...)

# S3 method for icd10cm
is_defined(x, short_code = guess_short(x),
  billable = FALSE, leaf = billable, ...)

# S3 method for icd10
is_defined(x, short_code = guess_short(x),
  billable = FALSE, leaf = billable, ...)

# S3 method for default
is_defined(x, short_code = guess_short(x), ...)

Arguments

x
vector if ICD codes to test whether defined in certain ICD code list

short_code
single logical value which determines whether the ICD-9 code provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this is guessed from the input data.

...
arguments passed on to other functions

billable
single logical value, identical to 'leaf'. Leaf is preferred as most adaptations of WHO ICD codes are not oriented around money.

leaf
single logical value, whether to limit return codes also by whether they are billable, i.e. leaf nodes. This is really only designed for use with ICD-9-CM, ICD-10-CM etc, since the WHO versions are not designed for billing, but for public health and death reporting.

leaf
New synonym for 'billable', which will be deprecated.

Value

logical vector

Details

The source data from CMS, and code to generate the R data are included in the source tree of this package.

Methods (by class)

icd9: Check whether ICD-9 code are defined. (ICD-9-CM by default)

icd10cm: Same for ICD-10-CM

icd10: Same for ICD-10, temporarily using ICD-10-CM until ICD-10 WHO is available in this package

default: default method which will guess the ICD version (9 vs 10, maybe WHO vs CM or other in the future) and dispatches again on that type.