Present-on-admission flags

See filter_poa for more details.

poa_choices

icd_poa_choices

Format

An object of class character of length 4.

Deprecated function names

Future versions of icd will drop the icd_ prefix. For example, charlson should be used in favor of icd_charlson. To distinguish icd function calls, consider using the prefix icd:: instead, e.g., icd::charlson. Functions which specifically operate on either ICD-9 or ICD-10 codes or their sub-types will retain the prefix. E.g. icd9_comorbid_ahrq. icd specific classes also retain the prefix, e.g., icd_wide_data.

Examples

  poa_choices

#> [1] "yes"    "no"     "notYes" "notNo"