Print a comorbidity map

The default is to summarize by printing the first seven comorbidities, and the first seven codes for each. To print the whole thing, just convert it to a list.

# S3 method for comorbidity_map
print(x, ..., n_comorbidities = 7, n_codes = 7)

Arguments

x
a list optionally with class comorbidity_map

...
further arguments are passed to print

n_comorbidities
single integer, number of comorbidities to print

n_codes
single integer, number of codes per comorbidity to print

Examples

icd9_map_ahrq

#> Showing first 7 comorbidities, and first 7 of each.

#> $CHF
#> [1] "39891" "40201" "40211" "40291" "40401" "40403" "40411"
#> 
#> $Valvular
#> [1] "0932"  "09320" "09321" "09322" "09323" "09324" "09325"
#> 
#> $PHTN
#> [1] "4151"  "41510" "41511" "41512" "41513" "41514" "41515"
#> 
#> $PVD
#> [1] "440"   "4400"  "44000" "44001" "44002" "44003" "44004"
#> 
#> $HTN
#> [1] "4011"  "40110" "40111" "40112" "40113" "40114" "40115"
#> 
#> $HTNcx
#> [1] "4010"  "40100" "40101" "40102" "40103" "40104" "40105"
#> 
#> $Paralysis
#> [1] "342"   "3420"  "34200" "34201" "34202" "34203" "34204"
#> 
#> ...

if (FALSE) {
print(icd9_map_ahrq)
print(icd9_map_ahrq, n_comorbidities = 3, n_codes = 3)
print(unclass(icd9_map_ahrq))
}