Refactor by integer matching levels in C++

Slightly slower for small factors, three times faster for one hundred million elements with two million new levels. Three times faster for any n > 1e6. With NA values, margin is smaller, but still beats base factor.

refactor(x, levels, na.rm = FALSE, exclude_na = TRUE,
  validate = FALSE)

Arguments

exclude_na
Simpler equivalent to base::factor exclude. By default, refactor will not count NA as a factor level if there are NA elements in the input data. As with base::factor, if exclude_na is TRUE, an NA level explicitly requested in the levels argument is still dropped.

validate
Single logical value, if TRUE, the input factor is checked for consistency. This should almost never be a problem (i.e., for any factors generated by R, or those carefully generated by the user). However, invalid factor input has the possibility of breaking assumptions in the C++ code. The default is FALSE. For huge data sets, and known valid input, skipping this check saves significant time. icd::factor_is_valid can be used independently.

Examples

# NOT RUN {
f <- factor(c(1, 2, 3))
icd:::refactor(f, c("2", "3"))
f <- factor(c(1, 2, NA))
icd:::refactor(f, c("2", "3", NA))
# }