ICD-9 chapters

icd9_chapters, icd9_chapters_sub and icd9_majors contain mappings from the higher level descriptions of ICD-9 codes to the ranges of ICD-9 codes they describe. Helpful in summarizing codes or grouping for human-readable output. These can easily be converted to a co-morbidity mapping, as shown in the vignette.

Format

list with chapter/sub-chapter or major names stored in list names, each with two element named character vector with start and end codes.

Source

http://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html

Details

001-139 Infectious And Parasitic Diseases

140-239 Neoplasms

240-279 Endocrine, Nutritional And Metabolic Diseases, And Immunity Disorders

280-289 Diseases Of The Blood And Blood-Forming Organs

290-319 Mental Disorders

320-389 Diseases Of The Nervous System And Sense Organs

390-459 Diseases Of The Circulatory System

460-519 Diseases Of The Respiratory System

520-579 Diseases Of The Digestive System

580-629 Diseases Of The Genitourinary System

630-679 Complications Of Pregnancy, Childbirth, And The Puerperium

680-709 Diseases Of The Skin And Subcutaneous Tissue

710-739 Diseases Of The Musculoskeletal System And Connective Tissue

740-759 Congenital Anomalies

760-779 Certain Conditions Originating In The Perinatal Period

780-799 Symptoms, Signs, And Ill-Defined Conditions

800-999 Injury And Poisoning

V01-V91 Supplementary Classification Of Factors Influencing Health Status And Contact With Health Services

E000-E999 Supplementary Classification Of External Causes Of Injury And Poisoning